refactor KendallTauSequenceDistance
